[SDKS-6577] Updated client version and change log

+++ client/src/main/java/io/split/client/impressions/strategy/ProcessImpressionOptimized.java
@@ -42,7 +42,7 @@ public class ProcessImpressionOptimized implements ProcessImpressionStrategy{
         }
         List<Impression> impressionForListener =  this._listenerEnabled ? impressions : null;
 
-        _telemetryRuntimeProducer.recordImpressionStats(ImpressionsDataTypeEnum.IMPRESSIONS_DEDUPED, impressions.size()-impressionsToQueue.size());
+        _telemetryRuntimeProducer.recordImpressionStats(ImpressionsDataTypeEnum.IMPRESSIONS_DEDUPED, impressions.size()- (long)impressionsToQueue.size());
 
         return new ImpressionsResult(impressionsToQueue, impressionForListener);
     }
